LayoutTests/fast/dom/shadow/elementfrompoint-expected.txt

PASS shadowRoot.elementFromPoint(x12, y12) is box12
PASS nestedShadowRoot.elementFromPoint(x22, y22) is box22
PASS shadowRoot.elementFromPoint(x22, y22) is nestedHost
PASS document.elementFromPoint(x22, y22) is host
PASS successfullyParsed is true

TEST COMPLETE
